backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.0
black==21.12b0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
cssselect==1.1.0
cssutils==2.3.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
emails==0.6
entrypoints==0.3
executing==0.8.2
fastapi==0.71.0
fastapi-login==1.7.3
fastapi-users==9.2.2
fastapi-users-db-opensearch @ git+https://github.com/KentonParton/fastapi-users-db-opensearch.git@78f7aff0480400bbeac01f12c17e04f1ae21fbed
frozenlist==1.3.0
great-expectations==0.15.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httpx==0.20.0
httpx-oauth==0.5.0
idna==3.3
importlib-metadata==4.10.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
lxml==4.7.1
makefun==1.13.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.22.1
opensearch-py==1.0.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
premailer==3.10.0
prometheus-client==0.13.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
stack-data==0.1.4
starlette==0.17.1
termcolor==1.1.0
terminado==0.13.0
testpath==0.5.0
tomli==1.2.3
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
uvicorn==0.17.0.post1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
yarl==1.7.2
zipp==3.7.0
